Extract object animation export to new module
