+++ io_scene_nif/modules/animation/animation_export.py
@@ -43,6 +43,7 @@ import mathutils
 from pyffi.formats.nif import NifFormat
 
 from io_scene_nif.modules import armature
+from io_scene_nif.modules.animation.object_export import ObjectAnimation
 from io_scene_nif.utility import nif_utils
 from io_scene_nif.utility.util_logging import NifLog
 from io_scene_nif.utility.util_global import NifOp
@@ -82,7 +83,7 @@ class Animation:
 
     def __init__(self, parent):
         self.nif_export = parent
-        self.object_animation = ObjectAnimation(parent)
+        self.object_animation = ObjectAnimation()
         self.material_animation = MaterialAnimation(parent)
         self.texture_animation = TextureAnimation(parent)
         self.fps = bpy.context.scene.render.fps
@@ -558,51 +559,3 @@ class MaterialAnimation:
             n_geom.add_controller(n_uv_ctrl)
 
 
-class ObjectAnimation:
-
-    def __init__(self, parent):
-        self.nif_export = parent
-
-    def export_object_vis_controller(self, n_node, b_obj):
-        """Export the visibility controller data."""
-
-        if not b_obj.animation_data and not b_obj.animation_data.action:
-            return
-        # get the hide fcurve
-        fcurves = [fcu for fcu in b_obj.animation_data.action.fcurves if "hide" in fcu.data_path]
-        if not fcurves:
-            return
-
-        # TODO [animation] which sort of controller should be exported?
-        #                  should this be driven by version number?
-        #                  we probably don't want both at the same time
-        # NiVisData = old style, NiBoolData = new style
-        n_vis_data = self.nif_export.objecthelper.create_block("NiVisData", fcurves)
-        n_bool_data = self.nif_export.objecthelper.create_block("NiBoolData", fcurves)
-
-        # we just leave interpolation at constant
-        n_bool_data.data.interpolation = NifFormat.KeyType.CONST_KEY
-        n_vis_data.num_keys = len(fcurves[0].keyframe_points)
-        n_vis_data.keys.update_size()
-        n_bool_data.data.num_keys = len(fcurves[0].keyframe_points)
-        n_bool_data.data.keys.update_size()
-        for b_point, n_vis_key, n_bool_key in zip(fcurves[0].keyframe_points, n_vis_data.keys, n_bool_data.data.keys):
-            # add each point of the curve
-            b_frame, b_value = b_point.co
-            n_vis_key.arg = n_bool_data.data.interpolation  # n_vis_data has no interpolation stored
-            n_vis_key.time = b_frame / bpy.context.scene.render.fps
-            n_vis_key.value = b_value
-            n_bool_key.arg = n_bool_data.data.interpolation
-            n_bool_key.time = n_vis_key.time
-            n_bool_key.value = n_vis_key.value
-        # if alpha data is present (check this by checking if times were added)
-        # then add the controller so it is exported
-        if fcurves[0].keyframe_points:
-            n_vis_ctrl = self.nif_export.objecthelper.create_block("NiVisController", fcurves)
-            n_vis_ipol = self.nif_export.objecthelper.create_block("NiBoolInterpolator", fcurves)
-            set_flags_and_timing(n_vis_ctrl, fcurves)
-            n_vis_ctrl.interpolator = n_vis_ipol
-            n_vis_ctrl.data = n_vis_data
-            n_vis_ipol.data = n_bool_data
-            # attach block to node
-            n_node.add_controller(n_vis_ctrl)
